requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.7
cycler==0.11.0
exceptiongroup==1.1.0
faiss-gpu==1.7.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py==3.7.0
hydra-core==1.3.1
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
kiwisolver==1.4.4
lightning-utilities==0.5.0
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.3
multidict==6.0.4
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.0
pickle5==0.0.11
Pillow==9.4.0
pluggy==1.0.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynndescent==0.5.8
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytorch-lightning==1.8.6
pytorch-metric-learning==2.0.1
PyYAML==6.0
record-keeper==0.9.32
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
umap-learn==0.5.3
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
